prov/efa: move rxr_info related function to efa_user_info.c
This patch moved the following functions from rxr_init.c to
efa_user_info.c

rxr_info_to_rxr() was renamed to efa_user_info_alter_rxr()
rxr_rdm_getinfo() was renamed to efa_user_info_get_rdm()
efa_getinfo()

The rest of rxr_init.c is purely related to rxr_env, hence
renamed this file to rxr_env.c.
