+++ prov/efa/Makefile.include
@@ -47,7 +47,7 @@ _efa_files = \
 	prov/efa/src/efa_prov_info.c \
 	prov/efa/src/efa_fork_support.c \
 	prov/efa/src/rxr/rxr_prov.c	\
-	prov/efa/src/rxr/rxr_init.c	\
+	prov/efa/src/rxr/rxr_env.c	\
 	prov/efa/src/rxr/rxr_cq.c	\
 	prov/efa/src/rxr/rxr_ep.c	\
 	prov/efa/src/rxr/rxr_cntr.c	\
@@ -74,7 +74,7 @@ _efa_headers = \
 	prov/efa/src/efa_prov_info.h \
 	prov/efa/src/efa_fork_support.h \
 	prov/efa/src/rxr/rxr.h \
-	prov/efa/src/rxr/rxr_init.h \
+	prov/efa/src/rxr/rxr_env.h \
 	prov/efa/src/rxr/rxr_cntr.h \
 	prov/efa/src/rxr/rxr_rma.h \
 	prov/efa/src/rxr/rxr_msg.h \
