+++ prov/efa/src/rxr/rxr_env.h
@@ -31,15 +31,57 @@
  * SOFTWARE.
  */
 
-#ifndef _RXR_INIT_H
-#define _RXR_INIT_H
+#ifndef _RXR_ENV_H
+#define _RXR_ENV_H
 
-int rxr_getinfo(uint32_t version, const char *node,
-		const char *service, uint64_t flags,
-		const struct fi_info *hints, struct fi_info **info);
+struct rxr_env {
+	int tx_min_credits;
+	int tx_queue_size;
+	int use_device_rdma;
+	int use_zcpy_rx;
+	int zcpy_rx_seed;
+	int enable_shm_transfer;
+	int shm_av_size;
+	int shm_max_medium_size;
+	int recvwin_size;
+	int ooo_pool_chunk_size;
+	int unexp_pool_chunk_size;
+	int readcopy_pool_size;
+	int atomrsp_pool_size;
+	int cq_size;
+	size_t max_memcpy_size;
+	size_t mtu_size;
+	size_t tx_size;
+	size_t rx_size;
+	size_t tx_iov_limit;
+	size_t rx_iov_limit;
+	int rx_copy_unexp;
+	int rx_copy_ooo;
+	int rnr_backoff_wait_time_cap; /* unit is us */
+	int rnr_backoff_initial_wait_time; /* unit is us */
+	size_t efa_cq_read_size;
+	size_t shm_cq_read_size;
+	size_t efa_max_medium_msg_size;
+	size_t efa_max_gdrcopy_msg_size;
+	size_t efa_min_read_msg_size;
+	size_t efa_min_read_write_size;
+	size_t efa_read_segment_size;
+	size_t efa_runt_size;
+	/* If first attempt to send a packet failed,
+	 * this value controls how many times firmware
+	 * retries the send before it report an RNR error
+	 * (via rdma-core error cq entry).
+	 *
+	 * The valid number is from
+	 *      0 (no retry)
+	 * to
+	 *      EFA_RNR_INFINITY_RETRY (retry infinitely)
+	 */
+	int rnr_retry;
+};
 
-void rxr_define_env();
+extern struct rxr_env rxr_env;
 
-void rxr_init_env();
+void rxr_env_initialize();
 
 #endif
